fix(core): 修复windows平台重复取消自启动的bug并优化UI显示
- 修复了在Windows平台上反复取消开机自启动时，由于注册表值不存在而引发的异常

- 调整了应用窗口的初始大小计算方式

- 优化了主界面和OBS设置界面的UI布局和样式

- 更新了前端依赖

+++ src/utils/tool.py
@@ -11,6 +11,9 @@ from pathlib import Path
 from ._version import __version__ as CURRENT_VERSION
 from src.notifypy import Notify
 
+if sys.platform == "win32":
+    import winreg
+
 logger = logging.getLogger("danmaku")
 
 
@@ -135,6 +138,20 @@ def get_autostart_command():
         return [f'"{sys.executable}"', f'"{os.path.abspath(sys.argv[0])}"']
 
 
+def check_registry_exists(app_name):
+    if sys.platform != "win32":
+        return False
+    try:
+        # import winreg
+        key = winreg.HKEY_CURRENT_USER
+        key_path = r"Software\Microsoft\Windows\CurrentVersion\Run"
+        with winreg.OpenKey(key, key_path, 0, winreg.KEY_READ) as run_key:
+            winreg.QueryValueEx(run_key, app_name)
+            return True
+    except FileNotFoundError:
+        return False
+
+
 def setup_autostart(enable: bool):
     """
     设置应用程序的自启动功能
@@ -149,7 +166,7 @@ def setup_autostart(enable: bool):
     command = get_autostart_command()
     try:
         if sys.platform == "win32":
-            import winreg
+            # import winreg
             key = winreg.HKEY_CURRENT_USER
             key_path = r"Software\Microsoft\Windows\CurrentVersion\Run"
             with winreg.OpenKey(key, key_path, 0, winreg.KEY_SET_VALUE) as run_key:
@@ -158,8 +175,9 @@ def setup_autostart(enable: bool):
                     winreg.SetValueEx(run_key, APP_NAME, 0, winreg.REG_SZ, cmd)
                     logger.info(f"Windows autostart enabled: {command}")
                 else:
-                    winreg.DeleteValue(run_key, APP_NAME)
-                    logger.info("Windows autostart disabled.")
+                    if check_registry_exists(APP_NAME):
+                        winreg.DeleteValue(run_key, APP_NAME)
+                        logger.info("Windows autostart disabled.")
                 return True
         elif sys.platform.startswith("linux"):
             user = getpass.getuser()
